+++ tests/testthat/test-software-versions.R
@@ -37,3 +37,10 @@ test_that("Testing software versions quarto output", {
   expect_equal(colnames(q), c("software", "version", "installed_version", "upgrade"))
   expect_true(all(is.logical(q$upgrade)))
 })
+
+test_that("Ensure that software versions are up to date", {
+  versions = get_latest_versions()
+  latest = create_software_tibble()
+  # If this test files, try running update_software_csv() first
+  expect_true(all(latest$version %in% versions$version))
+})